Added NoSoundSlimGuiTest
An abstract class for testing the GUI without sounds.

<feature, testing
