+++ src/test/java/com/github/fishio/gui/SlimGuiTest.java
@@ -6,7 +6,7 @@ import javafx.stage.Stage;
  * A GuiTest that only starts a "slim" version of the GUI: no window, no
  * FishIO instance, just JavaFX.
  */
-public class SlimGuiTest extends AppTest {
+public abstract class SlimGuiTest extends AppTest {
 	@Override
 	public void start(Stage stage) throws Exception { }
 }
